requirements.txt
torch
torchvision
timm
opencv-python
scikit-learn
Pillow